+++ src/hipay_enterprise/hipay_enterprise-17.php
@@ -226,6 +226,7 @@ class HipayEnterpriseNew extends Hipay_enterprise
             switch ($uxMode) {
                 case UXMode::DIRECT_POST:
                 case UXMode::HOSTED_FIELDS:
+                case UXMode::HOSTED_PAGE:
                     // set credit card for one click
                     $this->ccToken = new HipayCCToken($this);
                     $savedCC = $this->ccToken->getSavedCC($params['cart']->id_customer);
@@ -242,6 +243,8 @@ class HipayEnterpriseNew extends Hipay_enterprise
                             'activatedCreditCard' => array_keys($paymentProduct["products"]),
                             'confHipay' => $this->hipayConfigTool->getConfigHipay(),
                             'is_guest' => $this->customer->is_guest,
+                            'customerFirstName' => $this->customer->firstname,
+                            'customerLastName' => $this->customer->lastname
                         )
                     );
 
@@ -331,7 +334,7 @@ class HipayEnterpriseNew extends Hipay_enterprise
             );
             $this->context->controller->registerJavascript(
                 'hipay-sdk-js',
-                'https://libs.hipay.com/js/sdkjs.js',
+                $this->hipayConfigTool->getPaymentGlobal()['sdk_js_url'],
                 ['server' => 'remote', 'position' => 'bottom', 'priority' => 20]
             );
         }