+++ src/hipay_enterprise/upgrade/Upgrade-2.4.0.php
@@ -1,0 +1,332 @@
+<?php
+/**
+ * HiPay Enterprise SDK Prestashop
+ *
+ * 2017 HiPay
+ *
+ * NOTICE OF LICENSE
+ *
+ * @author    HiPay <support.tpp@hipay.com>
+ * @copyright 2017 HiPay
+ * @license   https://github.com/hipay/hipay-enterprise-sdk-prestashop/blob/master/LICENSE.md
+ */
+
+require_once(dirname(__FILE__) . '/../classes/helper/HipayDBQuery.php');
+
+function upgrade_module_2_4_0($module)
+{
+    $log = $module->getLogs();
+
+    $log->logInfos("Upgrade to 2.4.0");
+
+
+    $sql = 'ALTER TABLE `' . _DB_PREFIX_ . HipayDBQuery::HIPAY_CC_TOKEN_TABLE . '` 
+            CHANGE COLUMN `token` `token` VARCHAR(255) NOT NULL ;';
+
+
+    Db::getInstance()->execute($sql);
+
+    $keepParameters = array(
+        "visa" => array(
+            "currencies" => "",
+            "countries" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "frontPosition" => "",
+            "activated" => ""
+        ),
+        "mastercard" => array(
+            "currencies" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "countries" => "",
+            "frontPosition" => "",
+            "activated" => ""
+        ),
+        "cb" => array(
+            "currencies" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "countries" => "",
+            "frontPosition" => "",
+            "activated" => ""
+        ),
+        "maestro" => array(
+            "currencies" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "countries" => "",
+            "frontPosition" => "",
+            "activated" => ""
+        ),
+        "american-express" => array(
+            "currencies" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "countries" => "",
+            "frontPosition" => "",
+            "activated" => ""
+        ),
+        "bcmc" => array(
+            "currencies" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "countries" => "",
+            "frontPosition" => "",
+            "activated" => ""
+        ),
+        "3xcb" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "3xcb-no-fees" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "4xcb" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "4xcb-no-fees" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "4xcb-no-fees" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "aura" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "banamex" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "banco-do-brasil" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "bbva-bancomer" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "boleto-bancario" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "bradesco" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "caixa" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "itau" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "oxxo" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "santander-cash" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "santander-home-banking" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "dexia-directnet" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "bnpp-3xcb" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "bnpp-4xcb" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "giropay" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "ideal" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "ing-homepay" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "klarnainvoice" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "paypal" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "postfinance-card" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "postfinance-efinance" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "przelewy24" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "sdd" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "sisal" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "sofort-uberweisung" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => "",
+            "currencies" => "",
+            "countries" => ""
+        ),
+        "webmoney-transfert" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "yandex" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        ),
+        "carte-cadeau" => array(
+            "displayName" => "",
+            "frontPosition" => "",
+            "minAmount" => "",
+            "maxAmount" => "",
+            "activated" => ""
+        )
+    );
+
+    $module->hipayConfigTool->updateFromJSONFile($keepParameters);
+
+    return true;
+}